Refactor Ksql DB client helpers
