+++ docs/changes/20250721_progress.md
@@ -8,3 +8,6 @@
 - connect KsqlDslOptions.DlqTopicName to DlqOptions.TopicName via new DlqOptions property
 ## 2025-07-21 05:23 JST [assistant]
 - remove DefaultValueAttribute tests and fixed build
+## 2025-07-21 16:53 JST [codex]
+- refactored KsqlContext URL helpers to instance methods and added persistent HttpClient
+
