+++ src/Application/KsqlDbResponse.cs
@@ -1,3 +1,3 @@
-namespace Kafka.Ksql.Linq.Application;
+namespace Kafka.Ksql.Linq;
 
 public record KsqlDbResponse(bool IsSuccess, string Message, int? ErrorCode = null, string? ErrorDetail = null);
