--- src/Core/Context/KafkaContextOptions.cs
@@ -1,43 +1,0 @@
-using Kafka.Ksql.Linq.Configuration;
-using Kafka.Ksql.Linq.Core.Configuration;
-using Microsoft.Extensions.Configuration;
-
-namespace Kafka.Ksql.Linq.Core.Context;
-
-public class KafkaContextOptions
-{
-    /// <summary>
-    /// Common Kafka settings (BootstrapServers, ClientId, etc.)
-    /// </summary>
-    public CommonSection Common { get; set; } = new();
-
-    /// <summary>
-    /// Kafka bootstrap servers
-    /// </summary>
-    public string BootstrapServers
-    {
-        get => Common.BootstrapServers;
-        set => Common = new CommonSection { BootstrapServers = value };
-    }
-
-    /// <summary>
-    /// Schema Registry base URL (see docs_configuration_reference.md)
-    /// </summary>
-    public string SchemaRegistryUrl { get; set; } = string.Empty;
-
-    public ValidationMode ValidationMode { get; set; } = ValidationMode.Strict;
-
-    public IConfiguration? Configuration { get; set; }
-
-    public static KafkaContextOptions FromConfiguration(IConfiguration configuration)
-    {
-        return new KafkaContextOptions { Configuration = configuration };
-    }
-
-    public static KafkaContextOptions FromAppSettings(string path)
-    {
-        var configuration = new ConfigurationBuilder().AddJsonFile(path).Build();
-        return FromConfiguration(configuration);
-    }
-}
-
