+++ tests/Application/EventSetWithServicesRocksDbTests.cs
@@ -2,7 +2,6 @@ using Kafka.Ksql.Linq;
 using Kafka.Ksql.Linq.Application;
 using Kafka.Ksql.Linq.Configuration;
 using Kafka.Ksql.Linq.Core.Abstractions;
-using Kafka.Ksql.Linq.Core.Context;
 using Kafka.Ksql.Linq.Cache.Core;
 using Kafka.Ksql.Linq.Cache.Extensions;
 using Kafka.Ksql.Linq.Cache;
@@ -17,7 +16,7 @@ public class EventSetWithServicesRocksDbTests
 {
     private class TestContext : KsqlContext
     {
-        public TestContext() : base(new KafkaContextOptions()) { }
+        public TestContext() : base(new KsqlDslOptions()) { }
 
         protected override bool SkipSchemaRegistration => true;
