+++ tests/Application/KsqlContextAsyncTests.cs
@@ -1,7 +1,7 @@
 using Kafka.Ksql.Linq;
 using Kafka.Ksql.Linq.Application;
+using Kafka.Ksql.Linq.Configuration;
 using Kafka.Ksql.Linq.Core.Abstractions;
-using Kafka.Ksql.Linq.Core.Context;
 using System;
 using System.Reflection;
 using System.Threading.Tasks;
@@ -13,7 +13,7 @@ public class KsqlContextAsyncTests
 {
     private class TestContext : KsqlContext
     {
-        public TestContext() : base(new KafkaContextOptions()) { }
+        public TestContext() : base(new KsqlDslOptions()) { }
 
         protected override bool SkipSchemaRegistration => true;
     }
