+++ tests/Application/KsqlContextDlqSendTests.cs
@@ -6,7 +6,6 @@ using Kafka.Ksql.Linq.Core.Models;
 using Kafka.Ksql.Linq.Messaging.Producers;
 using Kafka.Ksql.Linq.Messaging.Producers.Core;
 using Kafka.Ksql.Linq.Core.Dlq;
-using Kafka.Ksql.Linq.Core.Context;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
@@ -39,7 +38,7 @@ public class KsqlContextDlqSendTests
 
     private class TestContext : KsqlContext
     {
-        public TestContext() : base(new KafkaContextOptions()) { }
+        public TestContext() : base(new KsqlDslOptions()) { }
 
         protected override bool SkipSchemaRegistration => true;
