+++ tests/Application/KsqlDbExecutionExtensionsTests.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Reflection;
 using Kafka.Ksql.Linq;
-using Kafka.Ksql.Linq.Core.Context;
 using Kafka.Ksql.Linq.Configuration;
 using Kafka.Ksql.Linq.Messaging.Configuration;
 using Kafka.Ksql.Linq.Core.Configuration;
@@ -15,7 +14,7 @@ public class KsqlDbExecutionExtensionsTests
 {
     private class DummyContext : KsqlContext
     {
-        public DummyContext(KafkaContextOptions opt) : base(opt) { }
+        public DummyContext(KsqlDslOptions opt) : base(opt) { }
         protected override bool SkipSchemaRegistration => true;
         protected override void OnModelCreating(Kafka.Ksql.Linq.Core.Abstractions.IModelBuilder modelBuilder) { }
     }
@@ -23,30 +22,30 @@ public class KsqlDbExecutionExtensionsTests
     [Fact]
     public void CreateClient_UsesSchemaRegistryHost()
     {
-        var ctx = new DummyContext(new KafkaContextOptions());
+        var ctx = new DummyContext(new KsqlDslOptions());
         var field = typeof(KsqlContext).GetField("_dslOptions", BindingFlags.NonPublic | BindingFlags.Instance)!;
         var opts = (KsqlDslOptions)field.GetValue(ctx)!;
         var schema = opts.SchemaRegistry;
         typeof(SchemaRegistrySection).GetProperty("Url")!.SetValue(schema, "http://example.com:8085");
 
         var method = typeof(KsqlContext)
-            .GetMethod("CreateClient", BindingFlags.NonPublic | BindingFlags.Static)!;
-        using var client = (HttpClient)method.Invoke(null, new object[] { ctx })!;
+            .GetMethod("CreateClient", BindingFlags.NonPublic | BindingFlags.Instance)!;
+        using var client = (HttpClient)method.Invoke(ctx, null)!;
         Assert.Equal(new Uri("http://example.com:8085"), client.BaseAddress);
     }
 
     [Fact]
     public void CreateClient_FallsBackToBootstrapServers()
     {
-        var ctx = new DummyContext(new KafkaContextOptions());
+        var ctx = new DummyContext(new KsqlDslOptions());
         var field = typeof(KsqlContext).GetField("_dslOptions", BindingFlags.NonPublic | BindingFlags.Instance)!;
         var opts = (KsqlDslOptions)field.GetValue(ctx)!;
         typeof(SchemaRegistrySection).GetProperty("Url")!.SetValue(opts.SchemaRegistry, "");
         typeof(CommonSection).GetProperty("BootstrapServers")!.SetValue(opts.Common, "example.com:9092");
 
         var method = typeof(KsqlContext)
-            .GetMethod("CreateClient", BindingFlags.NonPublic | BindingFlags.Static)!;
-        using var client = (HttpClient)method.Invoke(null, new object[] { ctx })!;
+            .GetMethod("CreateClient", BindingFlags.NonPublic | BindingFlags.Instance)!;
+        using var client = (HttpClient)method.Invoke(ctx, null)!;
         Assert.Equal(new Uri("http://example.com:9092"), client.BaseAddress);
     }
 }
