+++ tests/Mapping/AutomaticQueryFlowTests.cs
@@ -1,6 +1,5 @@
 using Kafka.Ksql.Linq;
 using Kafka.Ksql.Linq.Core.Abstractions;
-using Kafka.Ksql.Linq.Core.Context;
 using Kafka.Ksql.Linq.Core.Modeling;
 using Kafka.Ksql.Linq.Mapping;
 using Kafka.Ksql.Linq.Messaging.Abstractions;
@@ -44,7 +43,7 @@ public class AutomaticQueryFlowTests
 
     private class TestContext : KsqlContext
     {
-        public TestContext() : base(new KafkaContextOptions()) { }
+        public TestContext() : base(new KsqlDslOptions()) { }
         protected override bool SkipSchemaRegistration => true;
         public void SetProducer(object manager)
         {
