+++ tests/Query/QueryAnalyzerTests.cs
@@ -3,9 +3,9 @@ using System.Linq;
 using System.Linq.Expressions;
 using Kafka.Ksql.Linq.Query.Analysis;
 using Kafka.Ksql.Linq;
+using Kafka.Ksql.Linq.Configuration;
 using Kafka.Ksql.Linq.Application;
 using Kafka.Ksql.Linq.Core.Abstractions;
-using Kafka.Ksql.Linq.Core.Context;
 using Kafka.Ksql.Linq.Core.Modeling;
 using Xunit;
 
@@ -74,7 +74,7 @@ public class QueryAnalyzerTests
 
     private class TestContext : KsqlContext
     {
-        public TestContext() : base(new KafkaContextOptions()) { }
+        public TestContext() : base(new KsqlDslOptions()) { }
         protected override bool SkipSchemaRegistration => true;
 
         protected override void OnModelCreating(IModelBuilder modelBuilder)
